+++ jkorm/src/main/kotlin/com/tang/jkorm/constants/BaseMethodName.kt
@@ -55,6 +55,12 @@ object BaseMethodName {
         return method.name == SELECT_BY_ID && method.parameterCount == 1 && method.parameterTypes[0].name == "long"
     }
 
+    private const val COUNT = "count"
+
+    fun isCount(method: Method): Boolean {
+        return method.name == COUNT
+    }
+
     fun isBaseMethod(methodName: String): Boolean {
         return when (methodName) {
             INSERT, INSERT_SELECTIVE, UPDATE, UPDATE_SELECTIVE, DELETE, DELETE_BY_ID, SELECT, SELECT_BY_ID  -> true