+++ jkorm/src/main/kotlin/com/tang/jkorm/utils/ResultSetHandlers.kt
@@ -30,4 +30,9 @@ object ResultSetHandlers {
         return entity
     }
 
+    fun getCount(resultSet: ResultSet): Long {
+        resultSet.next()
+        return resultSet.getLong(1)
+    }
+
 }