+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 
  * Support function overloading
  * Support select condition query
+ * Support count query